@@ -1,6 +1,6 @@
 package ro.pub.cs.systems.eim.lab06.daytimeprotocol.views;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
